Have FillShaders all share the same texture used for intermediary rendering
